+++ docs/examples/analysis/plot_mcrals_kinetics.py
@@ -6,8 +6,8 @@
 # ======================================================================================
 # flake8: noqa
 """
-MCR-ALS optimization example with kinetic constraints
-=====================================================
+MCR-ALS with kinetic constraints
+================================
 
 In this example, we perform the MCR ALS optimization of the UV-vis of spectra resulting
 from a three-component reaction `A` \-> `B` \-> `C` which was investigated by UV–Vis
@@ -109,6 +109,9 @@ mcr_2.fit(X, Ckin)
 # Now, let\'s compare the concentration profile of MCR-ALS
 # (C = X(C$_{kin}^+$ X)$^+$) with
 # that of the optimized kinetic model (C$_{kin}$ \equiv$ `C_constrained`):
+
+# sphinx_gallery_thumbnail_number = 6
+
 _ = mcr_2.C.T.plot()
 _ = mcr_2.C_constrained.T.plot(clear=False)
 
@@ -119,8 +122,9 @@ _ = mcr_2.St.plot()
 _ = mcr_2.plotmerit(offset=0, nb_traces=10)
 
 # %%
-# This ends the example ! The following line can be uncommented if no plot shows when running
-# the .py script
+# This ends the example ! The following line can be uncommented if no plot shows when
+# running the .py script
 
 # %%
+
 # scp.show()