+++ docs/examples/analysis/plot_pca_iris.py
@@ -6,8 +6,8 @@
 # ======================================================================================
 # flake8: noqa
 """
-PCA analysis example
----------------------
+PCA example (iris dataset)
+--------------------------
 In this example, we perform the PCA dimensionality reduction of the classical `iris`
 dataset (Ronald A. Fisher.
 "The Use of Multiple Measurements in Taxonomic Problems. Annals of Eugenics, 7, pp.179-188, 1936).
@@ -89,8 +89,9 @@ ax = pca.scoreplot(scores, 1, 2, 3, color_mapping="labels")
 ax.view_init(10, 75)
 
 # %%
-# This ends the example ! The following line can be uncommented if no plot shows when running
-# the .py script
+# This ends the example ! The following line can be uncommented if no plot shows when
+# running the .py script
 
 # %%
+
 # scp.show()