+++ docs/examples/analysis/plot_pls.py
@@ -7,7 +7,7 @@
 # flake8: noqa
 """
 PLS regression example
--------------------------------
+----------------------
 In this example, we perform a PLS regression to predict the moisture of corn samples
 from their NIR spectra.
 
@@ -61,15 +61,16 @@ _ = pls.fit(X_train, y_train)
 # both train set and test set.
 
 # sphinx_gallery_thumbnail_number = 3
-ax = pls.parityplot(label="calibration")
+ax = pls.parityplot(label="calibration", s=150)
 _ = pls.parityplot(
-    y_test, pls.predict(X_test), c="red", label="validation", clear=False
+    y_test, pls.predict(X_test), s=150, c="red", label="validation", clear=False
 )
 _ = ax.legend(loc="lower right")
 
 # %%
-# This ends the example ! The following line can be uncommented if no plot shows when running
-# the .py script
+# This ends the example ! The following line can be uncommented if no plot shows when
+# running the .py script
 
 # %%
+
 # scp.show()