+++ spectrochempy/analysis/_base.py
@@ -158,11 +158,12 @@ class _set_output(object):
                             labels=["#%d" % (i) for i in range(X_transf.shape[0])],
                             title="components",
                         ),
-                        x=X.coord(-1),
+                        x=X.coord(-1).copy() if X.coord(-1) is not None else None,
                     )
                 if self.typex == "components":
                     X_transf.set_coordset(
-                        y=X.coord(0),  # cannot use X.y in case of transposed X
+                        y=X.coord(0).copy() if X.coord(0) is not None else None,
+                        # cannot use X.y in case of transposed X
                         x=Coord(
                             None,
                             labels=["#%d" % (i) for i in range(X_transf.shape[-1])],
@@ -176,7 +177,16 @@ class _set_output(object):
                             labels=["#%d" % (i) for i in range(X_transf.shape[-1])],
                             title="components",
                         ),
-                        x=X.coord(1),
+                        x=X.coord(1).copy() if X.coord(1) is not None else None,
+                    )
+                if self.typey == "features":
+                    X_transf.set_coordset(
+                        y=X.coord(1).copy() if X.coord(1) is not None else None,
+                        x=Coord(
+                            None,
+                            labels=["#%d" % (i) for i in range(X_transf.shape[-1])],
+                            title="components",
+                        ),
                     )
                 if self.typesingle == "components":
                     # occurs when the data are 1D such as ev_ratio...
@@ -655,6 +665,7 @@ class DecompositionAnalysis(AnalysisConfigurable):
     See Also
     --------
     EFA : Perform an Evolving Factor Analysis (forward and reverse).
+    FastICA : Perform Independent Component Analysis with a fast algorithm.
     IRIS : Integral inversion solver for spectroscopic data.
     MCRALS : Perform MCR-ALS of a dataset knowing the initial :math:`C` or :math:`S^T` matrix.
     NMF : Non-Negative Matrix Factorization.